consumed resources now takes a float that can indicate a proability of consuming one more resource

+++ mettagrid/src/metta/mettagrid/mettagrid_config.py
@@ -49,7 +49,7 @@ class ActionConfig(Config):
     enabled: bool = Field(default=True)
     # required_resources defaults to consumed_resources. Otherwise, should be a superset of consumed_resources.
     required_resources: dict[str, int] = Field(default_factory=dict)
-    consumed_resources: dict[str, int] = Field(default_factory=dict)
+    consumed_resources: dict[str, float] = Field(default_factory=dict)
 
 
 class AttackActionConfig(ActionConfig):
